+++ Makefile
@@ -35,7 +35,7 @@ endif
 REPO_ROOT				 = $(shell git rev-parse --show-toplevel)
 REVISION				?= $(shell git rev-parse --short HEAD)
 ACN_VERSION				?= $(shell git describe --exclude "azure-ipam*" --exclude "dropgz*" --exclude "zapai*" --tags --always)
-IPV6_HEALTHPROBE_TC_VERSION			?= $(notdir $(shell git describe --match "ipv6-healthprobe-tc*" --tags --always))
+IPV6_HP_BPF_VERSION			?= $(notdir $(shell git describe --match "ipv6-hp-bpf*" --tags --always))
 AZURE_IPAM_VERSION		?= $(notdir $(shell git describe --match "azure-ipam*" --tags --always))
 CNI_VERSION				?= $(ACN_VERSION)
 CNI_DROPGZ_VERSION		?= $(notdir $(shell git describe --match "dropgz*" --tags --always))
@@ -45,7 +45,7 @@ ZAPAI_VERSION			?= $(notdir $(shell git describe --match "zapai*" --tags --alway
 
 # Build directories.
 AZURE_IPAM_DIR = $(REPO_ROOT)/azure-ipam
-IPV6_HEALTHPROBE_TC_DIR = $(REPO_ROOT)/bpf-prog/ipv6-healthprobe-tc
+IPV6_HP_BPF_DIR = $(REPO_ROOT)/bpf-prog/ipv6-hp-bpf
 CNM_DIR = $(REPO_ROOT)/cnm/plugin
 CNI_NET_DIR = $(REPO_ROOT)/cni/network/plugin
 CNI_IPAM_DIR = $(REPO_ROOT)/cni/ipam/plugin
@@ -57,7 +57,7 @@ NPM_DIR = $(REPO_ROOT)/npm/cmd
 OUTPUT_DIR = $(REPO_ROOT)/output
 BUILD_DIR = $(OUTPUT_DIR)/$(GOOS)_$(GOARCH)
 AZURE_IPAM_BUILD_DIR = $(BUILD_DIR)/azure-ipam
-IPV6_HEALTHPROBE_TC_BUILD_DIR = $(BUILD_DIR)/bpf-prog/ipv6-healthprobe-tc
+IPV6_HP_BPF_BUILD_DIR = $(BUILD_DIR)/bpf-prog/ipv6-hp-bpf
 IMAGE_DIR  = $(OUTPUT_DIR)/images
 CNM_BUILD_DIR = $(BUILD_DIR)/cnm
 CNI_BUILD_DIR = $(BUILD_DIR)/cni
@@ -138,7 +138,7 @@ azure-cns: azure-cns-binary cns-archive
 acncli: acncli-binary acncli-archive
 azure-npm: azure-npm-binary npm-archive
 azure-ipam: azure-ipam-binary azure-ipam-archive
-ipv6-healthprobe-tc: ipv6-healthprobe-tc-binary ipv6-healthprobe-tc-archive
+ipv6-hp-bpf: ipv6-hp-bpf-binary ipv6-hp-bpf-archive
 
 
 ##@ Versioning
@@ -154,8 +154,8 @@ acncli-version: version
 azure-ipam-version: ## prints the azure-ipam version
 	@echo $(AZURE_IPAM_VERSION)
 
-ipv6-healthprobe-tc-version: ## prints the ipv6-healthprobe-tc version
-	@echo $(IPV6_HEALTHPROBE_TC_VERSION)
+ipv6-hp-bpf-version: ## prints the ipv6-hp-bpf version
+	@echo $(IPV6_HP_BPF_VERSION)
 
 cni-version: ## prints the cni version
 	@echo $(CNI_VERSION)
@@ -178,10 +178,10 @@ zapai-version: ## prints the zapai version
 azure-ipam-binary:
 	cd $(AZURE_IPAM_DIR) && CGO_ENABLED=0 go build -v -o $(AZURE_IPAM_BUILD_DIR)/azure-ipam$(EXE_EXT) -ldflags "-X github.com/Azure/azure-container-networking/azure-ipam/internal/buildinfo.Version=$(AZURE_IPAM_VERSION)" -gcflags="-dwarflocationlists=true"
 
-# Build the ipv6-healthprobe-tc binary.
-ipv6-healthprobe-tc-binary:
-	cd $(IPV6_HEALTHPROBE_TC_DIR) && CGO_ENABLED=0 go generate ./... 
-	cd $(IPV6_HEALTHPROBE_TC_DIR)/cmd/ipv6-healthprobe-tc && CGO_ENABLED=0 go build -v -o $(IPV6_HEALTHPROBE_TC_BUILD_DIR)$(EXE_EXT) -ldflags "-X main.version=$(IPV6_HEALTHPROBE_TC_VERSION)" -gcflags="-dwarflocationlists=true"
+# Build the ipv6-hp-bpf binary.
+ipv6-hp-bpf-binary:
+	cd $(IPV6_HP_BPF_DIR) && CGO_ENABLED=0 go generate ./... 
+	cd $(IPV6_HP_BPF_DIR)/cmd/ipv6-hp-bpf && CGO_ENABLED=0 go build -v -o $(IPV6_HP_BPF_BUILD_DIR)$(EXE_EXT) -ldflags "-X main.version=$(IPV6_HP_BPF_VERSION)" -gcflags="-dwarflocationlists=true"
 
 # Build the Azure CNM binary.
 cnm-binary:
@@ -256,7 +256,7 @@ endif
 ## Image name definitions.
 ACNCLI_IMAGE		= acncli
 AZURE_IPAM_IMAGE	= azure-ipam
-IPV6_HEALTHPROBE_TC_IMAGE		= ipv6-healthprobe-tc
+IPV6_HP_BPF_IMAGE		= ipv6-hp-bpf
 CNI_IMAGE			= azure-cni
 CNI_DROPGZ_IMAGE	= cni-dropgz
 CNS_IMAGE			= azure-cns
@@ -266,7 +266,7 @@ NPM_IMAGE			= azure-npm
 ACNCLI_PLATFORM_TAG				?= $(subst /,-,$(PLATFORM))$(if $(OS_VERSION),-$(OS_VERSION),)-$(ACN_VERSION)
 AZURE_IPAM_PLATFORM_TAG			?= $(subst /,-,$(PLATFORM))$(if $(OS_VERSION),-$(OS_VERSION),)-$(AZURE_IPAM_VERSION)
 AZURE_IPAM_WINDOWS_PLATFORM_TAG	?= $(subst /,-,$(PLATFORM))$(if $(OS_VERSION),-$(OS_VERSION),)-$(AZURE_IPAM_VERSION)-$(OS_SKU_WIN)
-IPV6_HEALTHPROBE_TC_IMAGE_PLATFORM_TAG		?= $(subst /,-,$(PLATFORM))$(if $(OS_VERSION),-$(OS_VERSION),)-$(IPV6_HEALTHPROBE_TC_VERSION)
+IPV6_HP_BPF_IMAGE_PLATFORM_TAG		?= $(subst /,-,$(PLATFORM))$(if $(OS_VERSION),-$(OS_VERSION),)-$(IPV6_HP_BPF_VERSION)
 CNI_PLATFORM_TAG				?= $(subst /,-,$(PLATFORM))$(if $(OS_VERSION),-$(OS_VERSION),)-$(CNI_VERSION)
 CNI_WINDOWS_PLATFORM_TAG		?= $(subst /,-,$(PLATFORM))$(if $(OS_VERSION),-$(OS_VERSION),)-$(CNI_VERSION)-$(OS_SKU_WIN)
 CNI_DROPGZ_PLATFORM_TAG 		?= $(subst /,-,$(PLATFORM))$(if $(OS_VERSION),-$(OS_VERSION),)-$(CNI_DROPGZ_VERSION)
@@ -374,34 +374,34 @@ azure-ipam-image-pull: ## pull azure-ipam container image.
 		IMAGE=$(AZURE_IPAM_IMAGE) \
 		TAG=$(AZURE_IPAM_PLATFORM_TAG)
 
-# ipv6-healthprobe-tc
+# ipv6-hp-bpf
 
-ipv6-healthprobe-tc-image-name: # util target to print the ipv6-healthprobe-tc image name.
-	@echo $(IPV6_HEALTHPROBE_TC_IMAGE)
+ipv6-hp-bpf-image-name: # util target to print the ipv6-hp-bpf image name.
+	@echo $(IPV6_HP_BPF_IMAGE)
 
-ipv6-healthprobe-tc-image-name-and-tag: # util target to print the ipv6-healthprobe-tc image name and tag.
-	@echo $(IMAGE_REGISTRY)/$(IPV6_HEALTHPROBE_TC_IMAGE):$(IPV6_HEALTHPROBE_TC_IMAGE_PLATFORM_TAG)
+ipv6-hp-bpf-image-name-and-tag: # util target to print the ipv6-hp-bpf image name and tag.
+	@echo $(IMAGE_REGISTRY)/$(IPV6_HP_BPF_IMAGE):$(IPV6_HP_BPF_IMAGE_PLATFORM_TAG)
 
-ipv6-healthprobe-tc-image: ## build ipv6-healthprobe-tc container image.
+ipv6-hp-bpf-image: ## build ipv6-hp-bpf container image.
 	$(MAKE) container \
-		DOCKERFILE=bpf-prog/ipv6-healthprobe-tc/$(OS).Dockerfile \
-		IMAGE=$(IPV6_HEALTHPROBE_TC_IMAGE) \
+		DOCKERFILE=bpf-prog/ipv6-hp-bpf/$(OS).Dockerfile \
+		IMAGE=$(IPV6_HP_BPF_IMAGE) \
 		EXTRA_BUILD_ARGS='--build-arg OS=$(OS) --build-arg ARCH=$(ARCH) --build-arg OS_VERSION=$(OS_VERSION) --build-arg DEBUG=$(DEBUG)'\
 		PLATFORM=$(PLATFORM) \
-		TAG=$(IPV6_HEALTHPROBE_TC_IMAGE_PLATFORM_TAG) \
+		TAG=$(IPV6_HP_BPF_IMAGE_PLATFORM_TAG) \
 		OS=$(OS) \
 		ARCH=$(ARCH) \
 		OS_VERSION=$(OS_VERSION)
 
-ipv6-healthprobe-tc-image-push: ## push ipv6-healthprobe-tc container image.
+ipv6-hp-bpf-image-push: ## push ipv6-hp-bpf container image.
 	$(MAKE) container-push \
-		IMAGE=$(IPV6_HEALTHPROBE_TC_IMAGE) \
-		TAG=$(IPV6_HEALTHPROBE_TC_IMAGE_PLATFORM_TAG)
+		IMAGE=$(IPV6_HP_BPF_IMAGE) \
+		TAG=$(IPV6_HP_BPF_IMAGE_PLATFORM_TAG)
 
-ipv6-healthprobe-tc-image-pull: ## pull ipv6-healthprobe-tc container image.
+ipv6-hp-bpf-image-pull: ## pull ipv6-hp-bpf container image.
 	$(MAKE) container-pull \
-		IMAGE=$(IPV6_HEALTHPROBE_TC_IMAGE) \
-		TAG=$(IPV6_HEALTHPROBE_TC_IMAGE_PLATFORM_TAG)
+		IMAGE=$(IPV6_HP_BPF_IMAGE) \
+		TAG=$(IPV6_HP_BPF_IMAGE_PLATFORM_TAG)
 
 # cni
 
@@ -772,12 +772,12 @@ ifeq ($(GOOS),linux)
 	cd $(AZURE_IPAM_BUILD_DIR) && $(ARCHIVE_CMD) $(AZURE_IPAM_ARCHIVE_NAME) azure-ipam$(EXE_EXT)
 endif
 
-# Create a ipv6-healthprobe-tc archive for the target platform.
-.PHONY: ipv6-healthprobe-tc-archive
-ipv6-healthprobe-tc-archive: ipv6-healthprobe-tc-binary
+# Create a ipv6-hp-bpf archive for the target platform.
+.PHONY: ipv6-hp-bpf-archive
+ipv6-hp-bpf-archive: ipv6-hp-bpf-binary
 ifeq ($(GOOS),linux)
-	$(MKDIR) $(IPV6_HEALTHPROBE_TC_BUILD_DIR)
-	cd $(IPV6_HEALTHPROBE_TC_BUILD_DIR) && $(ARCHIVE_CMD) $(IPV6_HEALTHPROBE_TC_ARCHIVE_NAME) ipv6-healthprobe-tc$(EXE_EXT)
+	$(MKDIR) $(IPV6_HP_BPF_BUILD_DIR)
+	cd $(IPV6_HP_BPF_BUILD_DIR) && $(ARCHIVE_CMD) $(IPV6_HP_BPF_ARCHIVE_NAME) ipv6-hp-bpf$(EXE_EXT)
 endif
 
 ##@ Utils
